BottinDetailsFragment: Create phone link independant on region
Linkify when doing addLinks() is dependant on the device's
default locale in order to format correctly the given text.
Some devices might be configured in locales different from
North America. In such a case, the phone number won't
have it's appropriate link as described here:

http://stackoverflow.com/a/40946996

Only verify if the phone number "can" be valid and create a
link for it. (Since we can't really validate the phone number
as such but all the numbers in the Directory should be valid
in theory)