AndroidCalendarManager: Define OWNER_ACCOUNT
As defined by the Android documentation,
OWNER_ACCOUNT must be defined when creating a
calendar. It was previously there but commented for
tests. More info on this available here:

https://developer.android.com/reference/android/provider/CalendarContract.Calendars.html

Let's place the owner account to the ETSMobile user. This
should fix some issues related to some calendar apps
crashing when trying to edit them (SPlanner on Samsung
devices for example)