TodayWidgetConfigureActivity: Add default values
If somebody previously configured a widget inside the application
and removes it afterward, he/she might want to reconfigure the
widget with the same attributes as before.

Since the preferences for a given widget are removed once it gets
deleted, add the required keys in order to save that said
preference.