TodayWidgetConfigureActivity: Remove constant values in getSharedPreferences()
It is recommended to use defined variables instead of constant
values because it might be possible that they change in possible
future revisions of Android. (Not that I believe that it will be
actually changed since SharedPreferences has been there for a bit)

In that case, let's do it for the good practices.