Replace Webviews by Chrome Custom Tabs
Using WebViews has advantages as it is a way to get native Webpage experience.
However, with it being somewhat coupled to the system, it might not be the
most secure/fast solution available.

Let's use the Chrome Tabs which are proven to be faster and be updated alongside
a user's compatible browser: (Mostly Chrome in that case)

- Add an Intent call to a web browser with an extra to a Custom tab if the said
device supports it.
- Remove BiblioFragment and MonETSFragment which aren't required anymore
- Replace "FRAGMENT" constants by "ITEM" as some items aren't technically
fragments anymore
- webservices.xml: Add the ETS library link as it was a constant initially.