MoodleAssignmentsActivity: Only show the bottom sheet if hidden
It looks like going between different state of a BottomSheetBehavior
makes it have some funky issues as far as the layout is concerned when
for example a user is to press a lot of time on a given assignment.

Show the bottom sheet only if it was hidden and hide it only when
the back button has been pressed.