Adapt RegistrationIntentService for recent post 7.0 changes
Starting in Android 8.0, (Oreo) background services have some
limitations in order to reduce the load on a device's system.
following are the final steps required to make FCM work in a
client's device.

- Make RegistrationIntentService extend JobIntentService which is
adapted to either make a background service run as a job or a service
based on the version of the device.
- Remove RegistrationIntentService calls in multiple activities. Only
execute it when authenticating the user and when a new token is
received from Firebase

Note: Solution for JobIntentService based on:
https://stackoverflow.com/a/49846410