Add runtime permissions for Moodle downloads
Since the target SDK has been changed, runtime permissions must be
handled correctly. Let's begin with Moodle that requires the storage
permissions to download a file to storage of the device.

If the said permission isn't granted, a message will prompt the user to
allow the wished permissions.

Add the runtime-permissions library from florent37 which simplifies
the flow.