Switch from runtime-permission to Permissions Dispatcher
The switch was made since the solution doesn't require reflection and
uses annotations which might be easier for one to see.

Also, don't request the READ_EXTERNAL_STORAGE permissions. Even though
the documentation recommends to not assume the permissions request for a
group, we'll take a risk with external storage ones being always
together.