HoraireFragment: Remove request for AppletsApiCalendar
RoboSpice is a library currently used for REST requests for the app.
However, it has been deprecated and it's use on Android 8.x+ can cause
crashes similar to this one:

https://github.com/stephanenicolas/robospice/issues/467

Since Notre-Dame is currently in development, let's just ensure that
the application is maintained without any hiccups. Let's start with
removing the parts that depends on a request that must be sent.