Trust ÉTS's root certificate for Signets network calls
Since ÉTS has changed it's website, it looks like there are probably some
intermediate certificate missing. Sadly, these certificates aren't
installed on older devices and thus can make them not connect correctly.

Fix this by including the certificate and including a custom trust and
keystore only for Signets calls. Create also the SignetsSSLTrust which
is a classic POJO for easier use for OkHttp3.

One caveat of this solution is that a new release should be made for
updating the certificate if necessary. However, this way, older devices
can still work.