Fix Activities overlap when logging in
Strangely after going to LoginActivity, there's a new Activity being but
on some newer versions of Android, it even goes as creating 2 activities
at the same time which is can be a little bothering for new users.

Remove the new Activity creation in LoginFragment and recreate the
Activity properly in MainActivity if the user is successfully logged in.