Refactor Signets AsyncTask
While it is understandable that the AsyncTask for a Signets request
is in DataManager, the class becomes almost unreadable in Android Studio
since the task itself is an object and that a warning for potential leaks
might occur.

Let's create a new class which will host the task and move
SignetsMethods as well from there.