SNS: Handle deletion of an device ARN endpoint
When a user is being logged off, the endpoint should be removed so there
a given application platform should be cleaner. In that case, we don't
need to configure the removal of the FCM token since the request goes
through AWS.

However, this doesn't prevent the endpoint from still being there when
the app will be uninstalled. In that case, the endpoint will most likely
be automatically disabled by AWS.