ETSFcmListerService: Create notification channel if not existent=
Starting from Android 8.0, applications needs a notification channel
based on the type of notifications received. FCM does create one by
default so let's use this one.

Since ÉTSMobile is in a maintenance state, only the default channel
is correct. More features should be implemented in Notre-Dame if that's
the case.